requirements.txt
tensorflow==2.1.0
keras==2.3.1
streamlit==0.70.0
Pillow==7.2.0
scipy==1.5.2
numpy==1.19.1
imageio==2.9.0
